Proof of Concept
This initial code collects keystrokes for the input box and outputs
to the console and a div while recording events in an array.
Upon clicking the Save Data button, the JSON formatted data is
sent to saveData.php where it is saved in a local file.
deleteData.php simply deletes the aforementioned local file.